Cherry-pick !4555, !4637, !4638, !4648, !4653, !4665, !4688, !4699, !4737, !4747, !4811 to 4.0-Release
Cherry-pick

Issue: #I8453T

+++ ecmascript/object_fast_operator.h
@@ -95,7 +95,7 @@ private:
     static inline JSTaggedValue AddPropertyByIndex(JSThread *thread, JSTaggedValue receiver, uint32_t index,
                                                    JSTaggedValue value);
 
-    static inline int32_t TryToElementsIndex(JSTaggedValue key);
+    static inline int64_t TryToElementsIndex(JSTaggedValue key);
 
     static inline bool GetNumFromString(const char *str, int len, int *index, int *num);
